requirements.txt
Django==4.1.2
djangorestframework==3.14.0
pygments==2.13.0
static3==0.7.0
pytest-django==4.5.2
pytest==7.2.0
python-dotenv==1.0.0
redis==4.5.5
django-redis==5.2.0
python-dateutil==2.8.2